requirements.txt
Django==2.2.10
django-filter==21.1
django-nested-admin==3.1.1
djangorestframework==3.8.2
psycopg2-binary==2.8
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
